Estimate dispersion parameters

Estimate dispersion parameters

estimate_dispersion(
  counts,
  manifold,
  cell.types,
  ages,
  model = "cCV",
  preprocess = TRUE,
  ncores = 1,
  size.factor = NULL,
  cell.types.to.use = NULL,
  ages.to.use = NULL,
  cell.types.cutoff = 10
)

Arguments

counts

An expression count matrix. The rows correspond to genes and the columns correspond to cells. It can be sparse.

manifold

The computed manifold from compute_manifold. It should be a matrix of the same dimension as counts.

cell.types

A vector with the same length as the columns of counts. Each value should be a string corresponding to the cell type of each cell in counts.

ages

A vector with the same length as the columns of counts. Each value should be a string corresponding to the age group of each cell in counts.

model

Model to be used for estimating dispersion parameters. Can be either "cCV" (constant CV), "cFF" (constant Fano factor), or "cVar" (constant variance).

preprocess

Whether the expression count matrix should be preprocessed, where all genes with zero counts will removed. Default is TRUE.

ncores

Number of cores to use. Default is 1.

size.factor

A vector of cell size normalization factors. Default uses mean library size normalization.

cell.types.to.use

A vector of unique cell types to be used in the estimation process. Default is all cell types.

ages.to.use

A vector of unique age groups to be used in the estimation process. Default is all age groups.

cell.types.cutoff

Minimum number of cells in a cell type x age group. Default is 10.

Value

A named list for all dispersion parameters. Each element is a table of dispersion parameters for a cell type. The columns in each table contain dispersion parameters for each age group.

Examples

# Loading test data from Tabula Muris Senis
devtools::install_github('fmicompbio/TabulaMurisSenisData')
#> Error in loadNamespace(x): there is no package called ‘devtools’
library(TabulaMurisSenisData)
#> Error in library(TabulaMurisSenisData): there is no package called ‘TabulaMurisSenisData’
tms_marrow <- TabulaMurisSenisDroplet(tissues = "Marrow")
#> Error in TabulaMurisSenisDroplet(tissues = "Marrow"): could not find function "TabulaMurisSenisDroplet"
tms_marrow_counts <- tms_marrow$Marrow@assays@data$counts
#> Error in eval(expr, envir, enclos): object 'tms_marrow' not found
rownames(tms_marrow_counts) <- rownames(tms_marrow$Marrow)
#> Error in eval(expr, envir, enclos): object 'tms_marrow' not found
colnames(tms_marrow_counts) <- colnames(tms_marrow$Marrow)
#> Error in eval(expr, envir, enclos): object 'tms_marrow' not found
tms_marrow_counts <- as(tms_marrow_counts, "dgCMatrix")
#> Error in eval(expr, envir, enclos): object 'tms_marrow_counts' not found

tms_marrow_ages <- which(tms_marrow$Marrow$age %in% c("3m", "30m"))
#> Error in eval(expr, envir, enclos): object 'tms_marrow' not found
tms_marrow_cell_types <- which(tms_marrow$Marrow$cell_ontology_class %in% c('erythroid progenitor',
'granulocytopoietic cell', 'hematopoietic precursor cell',
'megakaryocyte-erythroid progenitor cell', 'precursor B cell'))
#> Error in eval(expr, envir, enclos): object 'tms_marrow' not found
tms_marrow_counts <- tms_marrow_counts[, intersect(tms_marrow_ages, tms_marrow_cell_types)]
#> Error in eval(expr, envir, enclos): object 'tms_marrow_counts' not found

tms_marrow_manifold_neighbor <- compute_manifold(tms_marrow_counts, method = "neighbor")
#> Starting manifold computation...
#> Error in eval(expr, envir, enclos): object 'tms_marrow_counts' not found

tms_marrow_dispersions <- estimate_dispersion(tms_marrow_counts, tms_marrow_manifold_neighbor,
  tms_marrow$Marrow$cell_ontology_class[intersect(tms_marrow_ages, tms_marrow_cell_types)],
  tms_marrow$Marrow$age[intersect(tms_marrow_ages, tms_marrow_cell_types)],
  model = "cCV", ncores = 4, cell.types.cutoff = 10)
#> Preprocessing the count matrix
#> Error in eval(expr, envir, enclos): object 'tms_marrow_counts' not found